Added juggle_pal and bounceBlend
Added juggle_pal and bounceBlend and replaced ints in for loops with
uint8_t

+++ GFPM_LEDRoutines_helpers.ino
@@ -127,13 +127,13 @@ int lowestPoint() {
 
 
 void fadeall(uint8_t fade_all_speed) {
-  for (int i = 0; i < NUM_LEDS; i++) {
+  for (uint8_t i = 0; i < NUM_LEDS; i++) {
     leds[i].nscale8(fade_all_speed);
   }
 }
 
 void brightall(uint8_t bright_all_speed) {
-  for (int i = 0; i < NUM_LEDS; i++) {
+  for (uint8_t i = 0; i < NUM_LEDS; i++) {
     leds[i] += leds[i].scale8(bright_all_speed) ;
   }
 }
@@ -141,7 +141,7 @@ void brightall(uint8_t bright_all_speed) {
 
 void addGlitter( fract8 chanceOfGlitter)
 {
-  for ( int i = 0 ; i < 5 ; i++ ) {
+  for ( uint8_t i = 0 ; i < 5 ; i++ ) {
     if ( random8() < chanceOfGlitter) {
       leds[ random16(NUM_LEDS) ] += CRGB::White;
     }
